validators tests
TUI-1 Create a new API for query availability and booking of flights

+++ src/main/java/com/tui/proof/model/Paxes.java
@@ -11,9 +11,9 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
 @Data
 public class Paxes {
 
-    private short infants;
-    private short children;
-    private short adults;
+    private int infants;
+    private int children;
+    private int adults;
 
     @Override
     public String toString() {
